+++ src/Tests/Unit/Public/Get-TelegramStickerPackInfo.Tests.ps1
@@ -84,7 +84,7 @@ InModuleScope PoshGram {
                     StickerSetName = 'STPicard'
                 }
                 { Get-TelegramStickerPackInfo @getTelegramStickerPackInfoSplat
-                    Assert-MockCalled -CommandName Write-Warning -Times 1 -Scope It }
+                    Should -Invoke -CommandName Write-Warning -Times 1 -Scope It }
             } #it
 
             It 'should return the exception if the API returns an error' {