+++ src/Tests/Unit/Public/Send-TelegramContact.Tests.ps1
@@ -56,7 +56,7 @@ InModuleScope PoshGram {
                     ProtectContent      = $true
                 }
                 { Send-TelegramContact @sendTelegramContactSplat
-                    Assert-MockCalled -CommandName Write-Warning -Times 1 -Scope It }
+                    Should -Invoke -CommandName Write-Warning -Times 1 -Scope It }
             } #it
 
             It 'should return the exception if the API returns an error' {