+++ src/Tests/Unit/Public/Send-TelegramLocalSticker.Tests.ps1
@@ -113,7 +113,7 @@ InModuleScope PoshGram {
                     ErrorAction = 'SilentlyContinue'
                 }
                 { Send-TelegramLocalSticker @sendTelegramLocalStickerSplat
-                    Assert-MockCalled -CommandName Write-Warning -Times 1 -Scope It }
+                    Should -Invoke -CommandName Write-Warning -Times 1 -Scope It }
             } #it
 
             It 'should return the exception if the API returns an error' {