+++ src/Tests/Unit/Public/Send-TelegramLocation.Tests.ps1
@@ -53,7 +53,7 @@ InModuleScope PoshGram {
                     ErrorAction         = 'SilentlyContinue'
                 }
                 { Send-TelegramLocation @sendTelegramLocationSplat
-                    Assert-MockCalled -CommandName Write-Warning -Times 1 -Scope It }
+                    Should -Invoke -CommandName Write-Warning -Times 1 -Scope It }
             } #it
 
             It 'should return the exception if the API returns an error' {