+++ src/Tests/Unit/Public/Send-TelegramSticker.Tests.ps1
@@ -89,7 +89,7 @@ InModuleScope PoshGram {
                     ProtectContent      = $true
                 }
                 { Send-TelegramSticker @sendTelegramStickerSplat
-                    Assert-MockCalled -CommandName Write-Warning -Times 1 -Scope It }
+                    Should -Invoke -CommandName Write-Warning -Times 1 -Scope It }
             } #it
 
             It 'should return the exception if the API returns an error' {