+++ src/Tests/Unit/Public/Send-TelegramURLDocument.Tests.ps1
@@ -94,7 +94,7 @@ InModuleScope PoshGram {
                     ErrorAction         = 'SilentlyContinue'
                 }
                 { Send-TelegramURLDocument @sendTelegramURLDocumentSplat
-                    Assert-MockCalled -CommandName Write-Warning -Times 1 -Scope It }
+                    Should -Invoke -CommandName Write-Warning -Times 1 -Scope It }
             } #it
 
             It 'should return the exception if the API returns an error' {