+++ src/Tests/Unit/Public/Send-TelegramVenue.Tests.ps1
@@ -55,7 +55,7 @@ InModuleScope PoshGram {
                     ProtectContent      = $true
                 }
                 { Send-TelegramVenue @sendTelegramVenueSplat
-                    Assert-MockCalled -CommandName Write-Warning -Times 1 -Scope It }
+                    Should -Invoke -CommandName Write-Warning -Times 1 -Scope It }
             } #it
 
             It 'should return the exception if the API returns an error' {