+++ src/Tests/Unit/Public/Test-BotToken.Tests.ps1
@@ -42,7 +42,7 @@ InModuleScope PoshGram {
                     ErrorAction = 'SilentlyContinue'
                 }
                 { Test-BotToken @testBotTokenSplat
-                    Assert-MockCalled -CommandName Write-Warning -Times 1 -Scope It }
+                    Should -Invoke -CommandName Write-Warning -Times 1 -Scope It }
             } #it
 
             It 'should return the exception if the API returns an error' {